asset management system (basic) (#13)
* [FileIO] initial implementation of NativeFS and PhysFS

* [FileIO] manual revision of NativeFS and PhysFS

* [FileIO] re-organize and adding C++ wrapper

* [ams] adding ZipBuilder implementation using miniz

* wip: no idea on the design, simply some refactor

* wip: designing AMS interface (AssetManager)

* wip: asset manager interface design + implementation

* wip: os/fs path type clarification

* wip: simple variable name changes

* wip: application structure refactoring

* refactor editor program structure

* [ImGui] replacing default font with nerd font

* Rename plugins

* Vendor Find*.cmake target normalization

* wip: revamp build system + using justfile for shortcut

* [build] completely move to vcpkg manifest mode

* wip: update editor ui

* wip: update FSAPI design, need to update underlying fs plugins

* wip: update FSAPI design, updated ZipFS, PakFS, NativeFS, still need to update PhysFS

* wip: update FSAPI design (fixed all fs plugins, commit for progress)

* FSAPI: clean up NativeFS and PhysFS implementation

* FSAPI: revive editor before proceeding

* [AMS] adding KTX based texture assets

+++ Library/Lyra/RPIKit/FrameGraph.cpp
@@ -72,11 +72,11 @@ void FrameGraph::compile()
 
     // pass.refcnt++ for every resource write
     for (auto& pass : passes)
-        pass.refcnt = pass.writes.size();
+        pass.refcnt = static_cast<uint>(pass.writes.size());
 
     // resource.refcnt++ for every resource read
     for (auto& resource : resources)
-        resource.refcnt = resource.consumers.size();
+        resource.refcnt = static_cast<uint>(resource.consumers.size());
 
     // identify resources with refcnt == 0 and push them on a stack
     Stack<uint> unused_resources;
