+++ Plugins/Vulkan/VkPipeline.cpp
@@ -78,8 +78,8 @@ VulkanPipeline::VulkanPipeline(const GPURenderPipelineDescriptor& desc)
     auto viewport     = VkViewport{};
     viewport.x        = 0;
     viewport.y        = 0;
-    viewport.width    = extent.width;
-    viewport.height   = extent.height;
+    viewport.width    = static_cast<float>(extent.width);
+    viewport.height   = static_cast<float>(extent.height);
     viewport.minDepth = 0.0f;
     viewport.maxDepth = 1.0f;
