+++ TestKit/dynamic_uniform/main.cpp
@@ -63,7 +63,7 @@ struct DynamicUniformApp : public TestApp
 
         uint  width  = desc.width;
         uint  height = desc.height;
-        float fovy   = 1.05;
+        float fovy   = 1.05f;
         float aspect = float(width) / float(height);
         auto  proj   = glm::perspective(fovy, aspect, 0.1f, 100.0f);
         auto  view   = glm::lookAt(
@@ -164,7 +164,7 @@ struct DynamicUniformApp : public TestApp
         // commands
         command.resource_barrier(state_transition(backbuffer.texture, undefined_state(), color_attachment_state()));
         command.begin_render_pass(render_pass);
-        command.set_viewport(0, 0, desc.width, desc.height);
+        command.set_viewport(0, 0, static_cast<float>(desc.width), static_cast<float>(desc.height));
         command.set_scissor_rect(0, 0, desc.width, desc.height);
         command.set_pipeline(pipeline.pipeline);
         command.set_vertex_buffer(0, geometry.vbuffer);
