+++ TestKit/graphics_pipeline/main.cpp
@@ -53,7 +53,7 @@ struct GraphicsPipelineApp : public TestApp
     {
         uint  width  = desc.width;
         uint  height = desc.height;
-        float fovy   = 1.05;
+        float fovy   = 1.05f;
         float aspect = float(width) / float(height);
 
         geometry = Geometry::create_triangle();
@@ -136,7 +136,7 @@ struct GraphicsPipelineApp : public TestApp
         // commands
         command.resource_barrier(state_transition(backbuffer.texture, undefined_state(), color_attachment_state()));
         command.begin_render_pass(render_pass);
-        command.set_viewport(0, 0, desc.width, desc.height);
+        command.set_viewport(0, 0, static_cast<float>(desc.width), static_cast<float>(desc.height));
         command.set_scissor_rect(0, 0, desc.width, desc.height);
         command.set_pipeline(pipeline.pipeline);
         command.set_vertex_buffer(0, geometry.vbuffer);
